--- src/components/ranking/PassionUser/style.ts
@@ -1,63 +1,0 @@
-import { styled } from 'styled-components';
-
-export const Table = styled.table`
-  width: 100%;
-
-  font: var(--text-caption);
-  text-align: center;
-
-  & > :nth-child(12) {
-    margin-top: 20px;
-    padding: 3px 0;
-    border-radius: 4px;
-
-    background-color: var(--white);
-
-    font-weight: 500;
-  }
-`;
-
-export const Tbody = styled.tbody`
-  & > :nth-child(11) {
-    margin-top: 20px;
-    padding: 3px 0;
-    border-radius: 4px;
-
-    background-color: var(--white);
-
-    font-weight: 600;
-  }
-`;
-
-export const Tr = styled.tr`
-  display: grid;
-  grid-template-columns: 0.5fr 1.5fr 1fr 1fr 1fr;
-  align-items: center;
-`;
-
-export const Th = styled.th`
-  padding: 10px 0;
-
-  font: var(--text-body);
-  font-weight: 600;
-`;
-
-export const RankingTd = styled.td`
-  padding: 5px 0;
-  height: auto;
-
-  line-height: 0;
-`;
-
-export const Td = styled.td`
-  padding: 10px 0;
-`;
-
-export const LoadingSpinnerWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  height: 50px;
-  padding: 0;
-`;
